Refactor minimap to use log summary graphql api

+++ x-pack/plugins/infra/server/graphql/log_entries/schema.gql.ts
@@ -110,7 +110,7 @@ export const logEntriesSchema = gql`
       "The millisecond timestamp that corresponds to the end of the interval"
       end: Float!
       "The size of each bucket in milliseconds"
-      bucketSize: Int!
+      bucketSize: Float!
       "The query to filter the log entries by"
       filterQuery: String
     ): InfraLogSummaryInterval!
